+++ CHANGELOG.md
@@ -14,6 +14,7 @@
 - Fix range calculation on sheets not working anymore (#725)
 - Encumbrance applied incorrectly with Strength altering changes (#729)
 - Item aura strength was empty (#731)
+- Fix race condition causing issues with adding context notes and changes (#730)
 
 ### Changelog